meal-agent/requirements.txt
Flask
requests
google-cloud-secret-manager
google-cloud-aiplatform
Flask-CORS
